+++ test/mocha/helpers.js
@@ -1,34 +1,23 @@
 /*!
- * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
 const brAccount = require('bedrock-account');
 const brPassport = require('bedrock-passport');
 const database = require('bedrock-mongodb');
-const {promisify} = require('util');
 const sinon = require('sinon');
 const mockData = require('./mock.data');
 
 exports.stubPassport = async ({email = 'alpha@example.com'} = {}) => {
-  const actors = await exports.getActors(mockData);
   const passportStub = sinon.stub(brPassport, 'optionallyAuthenticated');
   passportStub.callsFake((req, res, next) => {
     req.user = {
-      account: mockData.accounts[email].account,
-      actor: actors[email],
+      account: mockData.accounts[email].account
     };
     next();
   });
   return passportStub;
-};
-
-exports.getActors = async mockData => {
-  const actors = {};
-  for(const [key, record] of Object.entries(mockData.accounts)) {
-    actors[key] = await brAccount.getCapabilities({id: record.account.id});
-  }
-  return actors;
 };
 
 exports.prepareDatabase = async mockData => {
@@ -45,7 +34,7 @@ exports.removeCollections = async (
     'profile-profileAgent',
     'profile-profileAgentCapabilitySet'
   ]) => {
-  await promisify(database.openCollections)(collectionNames);
+  await database.openCollections(collectionNames);
   for(const collectionName of collectionNames) {
     await database.collections[collectionName].deleteMany({});
   }
@@ -59,7 +48,7 @@ async function insertTestData(mockData) {
   for(const record of records) {
     try {
       await brAccount.insert(
-        {actor: null, account: record.account, meta: record.meta || {}});
+        {account: record.account, meta: record.meta || {}});
     } catch(e) {
       if(e.name === 'DuplicateError') {
         // duplicate error means test data is already loaded