+++ test/test.config.js
@@ -1,5 +1,5 @@
-/*
- * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2020-2022 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
@@ -7,7 +7,6 @@ const {config} = require('bedrock');
 const path = require('path');
 
 config.mocha.tests.push(path.join(__dirname, 'mocha'));
-const {permissions, roles} = config.permission;
 
 // Express
 config.express.useSession = true;
@@ -20,18 +19,6 @@ config.mongodb.dropCollections.collections = [];
 
 // allow self-signed certs in test framework
 config['https-agent'].rejectUnauthorized = false;
-
-// Account
-roles['bedrock-test.regular'] = {
-  id: 'bedrock-test.regular',
-  label: 'Account Test Role',
-  comment: 'Role for Test User',
-  sysPermission: [
-    permissions.ACCOUNT_ACCESS.id,
-    permissions.ACCOUNT_UPDATE.id,
-    permissions.ACCOUNT_INSERT.id
-  ]
-};
 
 // Profile
 config.profile.kms.baseUrl = `${config.server.baseUri}/kms`;